Added simpler session management with only session id being saved as a cookie and refreshing works now.

+++ ACSP-Backend/src/main/java/com/akjostudios/acsp/backend/services/security/SecurityService.java
@@ -86,7 +86,7 @@ public class SecurityService {
 		return keyPairGenerator.generateKeyPair();
 	}
 
-	public String generateToken(String sessionId, String userId, int seconds, RSAPrivateKey privateKey) {
+	public String generateToken(String sessionId, String userId, long seconds, RSAPrivateKey privateKey) {
 		Instant now = Instant.now();
 		Instant expiration = now.plusSeconds(seconds);
 		HashMap<String, Object> claims = new HashMap<>();
